Binding of Isaac: Afterbirth
Hello,
I currently admin along with 4 other people and group of 143 people and rising. We are always looking for more ways to increase in size and this looks like an amazing opportunity. Here is a link to the server in case you want to check it out. https://discord.gg/0gonaBVBNISU8zhS
Thanks,
Luke "Brightqwerty" Jacob