Resolve conflicting fixes relevant to case sensitive OSes.
I'd made bdStorage.json always capitalised as so to match the filename on
my system left by an earlier version. Whereas rauenzi's branch had made it
always bdstorage.json and the two changes, when merged, effectively
inverted the inconsistency. Fix it to use bdstorage.json per rauenzi but
to rename bdStorage.json if it exists (and bdstorage.json doesn't).