Feat/extend webpack and reactutils (#2007)
* Extend Webpack and ReactUtils

Adds a few webpack apis suggested by developers & myself and fixes some stuff

Adds an optional called `fatal` on all webpack searches which will throw when a module is not found.

Adds `getBulkObject` like `getBulk` but you pass a record and not an array

Adds `Filters.not` which just returns the opposite of the of the original filter

Adds `Filters.byComponentType`  which will search inside of memos and forwardRefs to get the type and pass it to your filter

Adds `ReactUtils.getType` which will search inside of memos/forwardRefs to get the type

Adds `ReactUtils.forceUpdateFiber` this will attempt to forceUpdate a fiber. If the fiber is a functional component, it must contain a state hook

Fixes a bug with getBulk where it wont exit early if all modules are found (unless any of the queries have `all` set to try)

Fixes a bug with `getLazy`/`waitForModule` for when signal is already aborted, it would ignore its aborted state and continue

Also changes `ReactUtils.wrapInHooks` to search inside of memo/forwardRef components, so the developer just has to pass the component type

* Didnt do the check in the right spot

* Add support for lazy components

* Comment out forceUpdateFiber

forceUpdateFiber does not work as intended

* Add options to getById

Adds fatal and raw to getById

* Fix getBulk problems

* Rename getBulkObject

Rename `getBulkObject` to `getBulkKeyed`

To match this https://github.com/tc39/proposal-await-dictionary